@@ -1,3 +1,5 @@
+using System.Text.Json.Serialization;
+
 namespace Community.Blazor.MapLibre.Models.Camera;
 
 public class FitBoundOptions : FlyToOptions
@@ -7,13 +9,21 @@ public class FitBoundOptions : FlyToOptions
     /// See those functions and {@link AnimationOptions} for information about options available.
     /// @defaultValue false
     /// </summary>
-    public bool Linear { get; set; } = false;
-    /// <summary>
-    /// The center of the given bounds relative to the map's center, measured in pixels.
-    /// </summary>
-    public new PointLike? Offset { get; set; }
+    [JsonPropertyName("linear")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public bool? Linear { get; set; }
+
     /// <summary>
     /// The maximum zoom level to allow when the map view transitions to the specified bounds.
     /// </summary>
+    [JsonPropertyName("maxZoom")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public double? MaxZoom { get; set; }
+
+    /// <summary>
+    /// The center of the given bounds relative to the map's center, measured in pixels.
+    /// </summary>
+    [JsonPropertyName("offset")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public new PointLike? Offset { get; set; }
 }
